Revert "Add --cabal-ghc/-pkg flags to 'configure'"
Undo these changes after a code review; 23Skidoo commented on them:

QUOTE
I don't think we need a special SetupWrapperFlags record - information
about the build compiler is not specific to setup scripts and could be
potentially used for other things.

The logical place to put this stuff is ConfigFlags and LocalBuildInfo. Add a
configBuildHcPath flag to ConfigFlags, and a buildCompiler field to
LocalBuildInfo. This will be consistent with already-existing configHcPath and
compiler.

Then you can just pass LocalBuildInfo.buildCompiler to setupWrapper via the
SetupScriptOptions.useCompiler field.

END-OF-QUOTE

See:
https://github.com/sheyll/cabal/commit/06c2501e48608024c1d80887c485c4596b8cf9e3